Added a way to create Julian dates using datetime objects.
Modified the jday function to accept either a datetime object or
a list of variables representing year, month, day, hour, minute, second; as before.

While making the old way a bit more of a hassle, datetime objects can more conveniently
be played around with and modified with such tools as timedelta.