Vectorize `tsince` in `propagation.sgp4()`
This speeds up propagation for large Time arrays in skyfield
`EarthSatellite.at()` method. No longer need to use for-loop for
`len(tsince) > 1`. Instead can use vectorized calculation.

WARNING: This is a work in progress.

The main change is to use `numpy` math functions instead of the built-in
`math` module, to enable element-wise calculations anywhere that
`tsince` is used.

The other change is all of the `if` checks on things like
`em` (eccentricity) and `inclm` (inclination) need a ndarray.any() call,
e.g. `if em >= 1.0` needs to change to `if (em >= 1.0).any()`
I've only updated a subset of those to enable testing of TLEs of certain
Flock and SkySat satellites, which are in LEO.