Vectorize `tsince` in `propagation.sgp4()`
This speeds up propagation for large Time arrays in skyfield
`EarthSatellite.at()` method. No longer need to use for-loop for
`len(tsince) > 1`. Instead can use vectorized calculation.

WARNING: This is a work in progress.

The main change is to use `numpy` math functions instead of the built-in
`math` module, to enable element-wise calculations anywhere that
`tsince` is used.

The other change is all of the `if` checks on things like
`em` (eccentricity) and `inclm` (inclination) need a ndarray.any() call,
e.g. `if em >= 1.0` needs to change to `if (em >= 1.0).any()`, and to
use `np.where` to replace certain `if-else` statements.

More work must be done for other methods within `sgp4.propagation`.
I've only touched `propagation.sgp4`. I imagine things will break if
sat.method = 'd' (functions related to deep space). A better way of
error handling for single time points within the array is necessary as
well.

I've only updated a subset of those to enable testing of TLEs of certain
Flock and SkySat satellites, which are in LEO.

I've ran some speed tests to show the benefit (~50X speed up).