Remove `@jit` decorator and numba import
Creating a lot of numpy arrays within a numba function is not a good
idea. However, `numpy` outperforms a `numba` function by an order of
magnitude for the case of `sgp4` looping over a large `tsince` array.

Thus, for a vectorized version of sgp4 using numpy arrays, it is best to
avoid using `@jit`.