Update propagation to fix mrt issue and add tests
The master SGP4 version is not returning `nan` when the mrt issue is
detected. This leads to a mismatch with some sats in `tcppver.out`.
This commit fixes this in the "vectorized" version, but not in master.

New tests were added to compare the master output to tcppver.out, which
uncovered the mrt issue.

New tests were added to compare "vectorized" output to tcppver.out.

New tests were added to compare "vectorized" output ot master output.

New tests were added to compare speed of "vectorized" (numpy) sgp4 to
master (jit) sgp4. The tests were aimed at understanding differences
when the length of the time array (tsince) is large (i.e. when
testing two week time period with time step of 1 minute).