Add a NumPy array routine to Satrec, too
It turns out that the routine doing (times x satellites) NumPy array
operations can, with only a tiny tweak, also do a loop over times for a
single satellite.  So, break out the routine into a helper and call it
from two places, once in a method for each object.