Turn implicit test numbers into explicit numbers
Instead of having the sample x,y,z values returned by SGP4 for various
gravity constants remain hidden behind code, let’s make them explicit.
This not only makes the tests easier for me to read, but means that if
we make adjustments later that change the actual numbers produced, we
will be warned with a test failure.