Remove test suite that is redundant with "tox"
This valiant attempt to test both the accelerated and non-accelerated
code path fails under at least two circumstances: first, if someone is
running the tests in a Python 2 only environment, this simply runs the
tests twice against the same Python code; and, second, if the C++ fails
to install or compile, then again this suite re-runs the tests on the
same Python code.

Since the “tox” setup guarantees testing of both code paths, let’s rely
on it as the One True mechanism for testing both the accelerated and the
slow code.  Now that the documentation draws attention to “tox” for
developers who might not recognize the significance of a “tox.ini”,
folks will hopefully start always using “tox” for running the tests.