Get gravity constants working with sgp4init()
This should have taken fewer hours that it did.  The main confounding
factor was that the `twoline2rv()` setup and `sgp4init()` setup paths
can currently produce very slightly different values for `jdsatepochF`,
making only a very small difference to the output, but unfortunately a
difference that’s significant compared to the difference in gravity
model.  It took quite a bit of hunting to realize that of all the floats
in the structure that could be different, that was the important one.
By invoking the SGP4 routine with `sgp4_tsince()`, the tests are now
independent of any issues with Julian date offsets.

As you can see from the comments, there were other snags; the most
memorable was a `char` that was being set with a full 4-byte value,
overwriting the automatic next to it.