Fix test failing locally: exporter AttributeError
Alas!  I had hoped that Travis CI Python 2.7 testing would fail exactly
when “tox” Python 2.7 fails locally, but the two are apparently not the
same, so I accepted this code from a PR when it in fact broke the tests
for me.  Here’s a fix that hopefully works in both cases.