Make 5 non-element attributes writeable for #65
These five Satrec attributes need to be writeable from Python if we want
to use Python for our new OMM import routines.  Since none of these
attributes are used when actually propagating the orbit, there should
not be any danger of these setters letting a satellite get into an
inconsistent state (which is why the orbital parameters themselves are
read-only and must be set through `sgp4init()`).