Add a note about SGP4 ignoring ndot and nddot
If you skip past parts of the package documentation and on to the __Providing your own elements_ section you'll have trouble understanding some of the parameters, particularly the `ndot` and `nddot` ones, which, according to other parts of the documentation, are ignored by SGP4. To make getting started with this API easier for newbies (i.e., me), I've added a small note that these two parameters are ignored by SGP4.
I've also added named parameters which makes it easier to use the example code in projects.