Document the complete list of supported attributes
This also, happily, removes `setup.py` code that was pulling in an
attribute list from an old docstring, a trick which was forever causing
me disappointment when I’d open `__init__.py` to search its docstring
for a particular attribute that I was then never able to find.

This took several hours to compile, including a series of test scripts
to verify units, but was well worth it for the discoveries along the way.