Have `test.sh` test both compiled and pure Python
This replaces the `SGP4_FORCE_COMPILE` environment variable, which was
merely binary, with a `PYTHON_SGP4_COMPILE` variable that can not only
tell `setup.py` to insist on compiling, but can also turn it off.

So we now, for the first time, locally test three cases: Python 2, and
Python 3 with compilation both on and off.